+++ agents-manage-api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @inkeep/agents-manage-api
 
+## 0.32.2
+
+### Patch Changes
+
+- c228770: update create-agents setup script
+- Updated dependencies [c228770]
+  - @inkeep/agents-core@0.32.2
+
+## 0.32.1
+
+### Patch Changes
+
+- 5bd3d93: update dev deps agent-core
+- Updated dependencies [5bd3d93]
+  - @inkeep/agents-core@0.32.1
+
 ## 0.32.0
 
 ### Minor Changes